--- backend/src/main/java/com/votogether/domain/vote/repository/VoteCustomRepository.java
@@ -1,10 +1,0 @@
-package com.votogether.domain.vote.repository;
-
-import com.votogether.domain.vote.dto.VoteStatus;
-import java.util.List;
-
-public interface VoteCustomRepository {
-
-    List<VoteStatus> findVoteCountByPostOptionIdGroupByAgeRangeAndGender(final Long postOptionId);
-
-}
