+++ backend/src/test/java/com/votogether/RepositoryTest.java
@@ -1,7 +1,6 @@
 package com.votogether;
 
 import com.votogether.config.JpaConfig;
-import com.votogether.config.QuerydslConfig;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -12,7 +11,7 @@ import org.springframework.context.annotation.Import;
 
 @Target(value = ElementType.TYPE)
 @Retention(value = RetentionPolicy.RUNTIME)
-@Import({JpaConfig.class, QuerydslConfig.class})
+@Import(JpaConfig.class)
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DataJpaTest
 public @interface RepositoryTest {
